feat: AuthCredential added, service documented, dockerfiles renamed, doc added, unused files removed
