+++ README.md
@@ -1,1 +1,3 @@
 # zz2_ws_backend
+
+[Notes-bgll](./note.md)
