+++ turradgiver-api/Controllers/AuthController.cs
@@ -1,14 +1,13 @@
 using DAL.Models;
 using DAL.Repositories;
-using turradgiver_api.Dtos.UserSignUpDto;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using turradgiver_api.Dtos.UserSignInDto;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using turradgiver_api.Services;
 using System.Threading.Tasks;
 using turradgiver_api.Utils;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
+using turradgiver_api.Dtos.Auth;
 
 namespace turradgiver_api.Controllers
 {
@@ -16,11 +15,8 @@ namespace turradgiver_api.Controllers
     [Route("[controller]")]
     public class AuthController : ControllerBase
     {
-
         private readonly ILogger<AuthController> _logger;
-
         private readonly IRepository<User> _userRepo;
-        // private readonly IGenericRepository<User> _repo;
         private readonly IAuthService _authService;
 
         public AuthController(ILogger<AuthController> logger,IAuthService authService, IRepository<User> userRepo)
@@ -31,7 +27,7 @@ namespace turradgiver_api.Controllers
         }
 
         [Authorize]
-        [HttpGet("getAll")]
+        [HttpGet("get-all")]
         public IActionResult GetAll()
         {
             // _logger.LogInformation("Object",o);
@@ -42,7 +38,7 @@ namespace turradgiver_api.Controllers
         public async Task<IActionResult> SignUp([FromBody] UserSignUpDto userSignUpDto)
         {
             _logger.LogInformation("UserSignUpDto",userSignUpDto);
-            Response<string> res = await _authService.Register(new User(userSignUpDto.Username, userSignUpDto.Email), userSignUpDto.Password);
+            Response<AuthCredential> res = await _authService.Register(new User(userSignUpDto.Username, userSignUpDto.Email), userSignUpDto.Password);
             if (! res.Success){
                 return Unauthorized(res.Message);
             }
@@ -56,15 +52,11 @@ namespace turradgiver_api.Controllers
             {
                 BadRequest();
             }
-            Response<string> res = await _authService.Login(userSignInDto.Email,userSignInDto.Password);
+            Response<AuthCredential> res = await _authService.Login(userSignInDto.Email,userSignInDto.Password);
              if (! res.Success){
                 return BadRequest(res.Message);
             }
-            return Ok(res);
-            
-            
+            return Ok(res);   
         }
-
-
     }
 }
